rename plugin name from devices to deviceshare
